+++ core/src/main/java/me/wolfyscript/utilities/util/eval/operators/ComparisonOperatorGreaterEqual.java
@@ -16,29 +16,22 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.util.value_providers;
+package me.wolfyscript.utilities.util.eval.operators;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import me.wolfyscript.utilities.util.NamespacedKey;
-import me.wolfyscript.utilities.util.context.EvalContext;
+import me.wolfyscript.utilities.util.eval.context.EvalContext;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProvider;
 
-public abstract class ValueProviderVariable<V> extends AbstractValueProvider<V> {
+public class ComparisonOperatorGreaterEqual<V extends Comparable<V>> extends ComparisonOperator<V> {
 
-    @JsonProperty("var")
-    private final String variable;
-    @JsonIgnore
-    private final Class<V> typeClass;
+    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("greater_equal");
 
-    protected ValueProviderVariable(NamespacedKey key, Class<V> typeClass, String variable) {
-        super(key);
-        this.typeClass = typeClass;
-        this.variable = variable;
+    protected ComparisonOperatorGreaterEqual(ValueProvider<V> thisValue, ValueProvider<V> thatValue) {
+        super(KEY, thisValue, thatValue);
     }
 
     @Override
-    public V getValue(EvalContext context) {
-        return typeClass.cast(context.getVariable(variable));
+    public boolean evaluate(EvalContext context) {
+        return this.thisValue.getValue(context).compareTo(this.thatValue.getValue(context)) >= 0;
     }
-
 }
