+++ core/src/main/java/me/wolfyscript/utilities/util/eval/operators/ComparisonOperatorLess.java
@@ -16,17 +16,22 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.util.value_providers;
+package me.wolfyscript.utilities.util.eval.operators;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
 import me.wolfyscript.utilities.util.NamespacedKey;
+import me.wolfyscript.utilities.util.eval.context.EvalContext;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProvider;
 
-public class ValueProviderIntegerVar extends ValueProviderVariable<Integer> implements ValueProviderInteger {
+public class ComparisonOperatorLess<V extends Comparable<V>> extends ComparisonOperator<V> {
 
-    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("int/var");
+    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("less");
 
-    public ValueProviderIntegerVar(@JsonProperty("var") String variable) {
-        super(KEY, Integer.class, variable);
+    protected ComparisonOperatorLess(ValueProvider<V> thisValue, ValueProvider<V> thatValue) {
+        super(KEY, thisValue, thatValue);
     }
 
+    @Override
+    public boolean evaluate(EvalContext context) {
+        return this.thisValue.getValue(context).compareTo(this.thatValue.getValue(context)) < 0;
+    }
 }
