+++ core/src/main/java/me/wolfyscript/utilities/util/eval/value_providers/ValueProviderConditioned.java
@@ -16,15 +16,13 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.util.value_providers;
+package me.wolfyscript.utilities.util.eval.value_providers;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import me.wolfyscript.utilities.util.NamespacedKey;
-import me.wolfyscript.utilities.util.context.EvalContext;
-import me.wolfyscript.utilities.util.operators.BoolOperator;
-import me.wolfyscript.utilities.util.operators.Operator;
-import me.wolfyscript.utilities.util.value_comparators.ValueComparator;
+import me.wolfyscript.utilities.util.eval.context.EvalContext;
+import me.wolfyscript.utilities.util.eval.operators.BoolOperator;
 
 public class ValueProviderConditioned<V> extends AbstractValueProvider<V> {
