+++ core/src/main/java/me/wolfyscript/utilities/util/eval/value_providers/ValueProviderFloat.java
@@ -16,27 +16,11 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.util.value_providers;
+package me.wolfyscript.utilities.util.eval.value_providers;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import me.wolfyscript.utilities.util.NamespacedKey;
-import me.wolfyscript.utilities.util.context.EvalContext;
+import me.wolfyscript.utilities.util.eval.context.EvalContext;
 
-public class ValueProviderIntegerConst extends AbstractValueProvider<Integer> implements ValueProviderInteger {
+public interface ValueProviderFloat extends ValueProvider<Float> {
 
-    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("int/const");
-
-    private final int value;
-
-    @JsonCreator
-    public ValueProviderIntegerConst(@JsonProperty("value") int value) {
-        super(KEY);
-        this.value = value;
-    }
-
-    @Override
-    public Integer getValue(EvalContext context) {
-        return value;
-    }
+    Float getValue(EvalContext context);
 }
