+++ core/src/main/java/me/wolfyscript/utilities/util/eval/value_providers/ValueProviderFloatConst.java
@@ -16,25 +16,32 @@
  *     along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package me.wolfyscript.utilities.util.operators;
+package me.wolfyscript.utilities.util.eval.value_providers;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import me.wolfyscript.utilities.util.NamespacedKey;
-import me.wolfyscript.utilities.util.context.EvalContext;
-import me.wolfyscript.utilities.util.value_providers.ValueProvider;
+import me.wolfyscript.utilities.util.eval.context.EvalContext;
 
-public class ComparisonOperatorEqual<V extends Comparable<V>> extends ComparisonOperator<V> {
+public class ValueProviderFloatConst extends AbstractValueProvider<Float> implements ValueProviderFloat {
 
-    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("equal");
+    public static final NamespacedKey KEY = NamespacedKey.wolfyutilties("float/const");
+
+    private final float value;
 
     @JsonCreator
-    protected ComparisonOperatorEqual(@JsonProperty("this") ValueProvider<V> thisValue, @JsonProperty("that") ValueProvider<V> thatValue) {
-        super(KEY, thisValue, thatValue);
+    public ValueProviderFloatConst(@JsonProperty("value") float value) {
+        super(KEY);
+        this.value = value;
     }
 
     @Override
-    public boolean evaluate(EvalContext context) {
-        return this.thisValue.getValue(context).compareTo(this.thatValue.getValue(context)) == 0;
+    public Float getValue(EvalContext context) {
+        return value;
+    }
+
+    @Override
+    public NamespacedKey getNamespacedKey() {
+        return null;
     }
 }
