+++ wolfyutilities/src/main/java/me/wolfyscript/utilities/main/WUPlugin.java
@@ -86,15 +86,15 @@ import me.wolfyscript.utilities.util.json.jackson.serialization.LocationSerializ
 import me.wolfyscript.utilities.util.json.jackson.serialization.PotionEffectSerialization;
 import me.wolfyscript.utilities.util.json.jackson.serialization.PotionEffectTypeSerialization;
 import me.wolfyscript.utilities.util.json.jackson.serialization.VectorSerialization;
-import me.wolfyscript.utilities.util.operators.ComparisonOperatorEqual;
-import me.wolfyscript.utilities.util.operators.ComparisonOperatorGreater;
-import me.wolfyscript.utilities.util.operators.ComparisonOperatorGreaterEqual;
-import me.wolfyscript.utilities.util.operators.ComparisonOperatorLess;
-import me.wolfyscript.utilities.util.operators.ComparisonOperatorLessEqual;
-import me.wolfyscript.utilities.util.operators.LogicalOperatorAnd;
-import me.wolfyscript.utilities.util.operators.LogicalOperatorNot;
-import me.wolfyscript.utilities.util.operators.LogicalOperatorOr;
-import me.wolfyscript.utilities.util.operators.Operator;
+import me.wolfyscript.utilities.util.eval.operators.ComparisonOperatorEqual;
+import me.wolfyscript.utilities.util.eval.operators.ComparisonOperatorGreater;
+import me.wolfyscript.utilities.util.eval.operators.ComparisonOperatorGreaterEqual;
+import me.wolfyscript.utilities.util.eval.operators.ComparisonOperatorLess;
+import me.wolfyscript.utilities.util.eval.operators.ComparisonOperatorLessEqual;
+import me.wolfyscript.utilities.util.eval.operators.LogicalOperatorAnd;
+import me.wolfyscript.utilities.util.eval.operators.LogicalOperatorNot;
+import me.wolfyscript.utilities.util.eval.operators.LogicalOperatorOr;
+import me.wolfyscript.utilities.util.eval.operators.Operator;
 import me.wolfyscript.utilities.util.particles.animators.Animator;
 import me.wolfyscript.utilities.util.particles.animators.AnimatorBasic;
 import me.wolfyscript.utilities.util.particles.animators.AnimatorCircle;
@@ -113,15 +113,14 @@ import me.wolfyscript.utilities.util.particles.timer.Timer;
 import me.wolfyscript.utilities.util.particles.timer.TimerLinear;
 import me.wolfyscript.utilities.util.particles.timer.TimerPi;
 import me.wolfyscript.utilities.util.particles.timer.TimerRandom;
-import me.wolfyscript.utilities.util.value_providers.ValueProvider;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderConditioned;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderFloatConst;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderFloatVar;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderIntegerConst;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderIntegerVar;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderStringConst;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderStringVar;
-import me.wolfyscript.utilities.util.value_providers.ValueProviderVariable;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProvider;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderConditioned;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderFloatConst;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderFloatVar;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderIntegerConst;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderIntegerVar;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderStringConst;
+import me.wolfyscript.utilities.util.eval.value_providers.ValueProviderStringVar;
 import me.wolfyscript.utilities.util.version.ServerVersion;
 import me.wolfyscript.utilities.util.world.WorldUtils;
 import net.kyori.adventure.platform.bukkit.BukkitAudiences;
